Further divided UsersListComponent into components. Renamed some file names so they can be read more easily.
