+++ MedicalDocManagment.WebUI/MedicalDocManagment.WebUI.csproj
@@ -139,7 +139,8 @@
   </ItemGroup>
   <ItemGroup>
     <Content Include="app\views\userslist-component.html" />
-    <Content Include="app\views\userslistitem-component.html" />
+    <Content Include="app\views\userslist-item-component.html" />
+    <Content Include="app\views\userslist-item-actionlist-component.html" />
     <Content Include="Content\bootstrap.css" />
     <Content Include="Content\bootstrap.min.css" />
     <Content Include="Content\Site.css" />
@@ -200,8 +201,9 @@
     <TypeScriptCompile Include="app\models\rolemodel.ts" />
     <TypeScriptCompile Include="app\models\usermodel.ts" />
     <TypeScriptCompile Include="app\userslist-component.ts" />
-    <TypeScriptCompile Include="app\userslistitem-component.ts" />
-    <TypeScriptCompile Include="app\userslistpagination-component.ts" />
+    <TypeScriptCompile Include="app\userslist-item-component.ts" />
+    <TypeScriptCompile Include="app\userslist-item-actionlist-component.ts" />
+    <TypeScriptCompile Include="app\userslist-pagination-component.ts" />
   </ItemGroup>
   <PropertyGroup>
     <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">10.0</VisualStudioVersion>
