+++ MedicalDocManagment.WebUI/app/userslist-item-actionlist-component.ts
@@ -1,0 +1,13 @@
+﻿import { Component, Input } from '@angular/core';
+import { Response } from '@angular/http';
+
+import UserModel from './models/usermodel'
+
+@Component({
+    selector: '[usersListItemActionList]',
+    templateUrl: './app/views/userslist-item-actionlist-component.html'
+})
+
+export class UsersListItemActionListComponent {
+}
+
