+++ MedicalDocManagment.WebUI/app/views/userslist-component.html
@@ -7,6 +7,7 @@
         <th>E-mail</th>
         <th>Посада</th>
         <th>Роль</th>
+        <th>Дії над користувачем</th>
     </tr>
     <tr *ngFor="let user of users" usersListItem [user]="user"></tr>
 </table>
