+++ MedicalDocManagment.WebUI/app/views/userslist-item-component.html
@@ -1,0 +1,6 @@
+﻿<td>{{user.username}}</td>
+<td>{{user.avatar}}</td>
+<td>{{user.email}}</td>
+<td>{{user.position}}</td>
+<td>{{user.role.role}}</td>
+<td usersListItemActionList></td>
