--- MedicalDocManagment.WebUI/app/views/userslistitem-component.html
@@ -1,6 +1,0 @@
-﻿
-<td>{{user.username}}</td>
-<td>{{user.avatar}}</td>
-<td>{{user.email}}</td>
-<td>{{user.position}}</td>
-<td>{{user.role.role}}</td>
